requirements.txt
httpx
cryptography
onnxruntime
opencv-python
numpy
torch
torchvision
Pillow
matplotlib
tqdm
shutil